fix: enable depcheck plugin, fees tolong, update dep-sha256.json

+++ src/test/kotlin/it/pagopa/ecommerce/eventdispatcher/services/v2/NodeServiceTests.kt
@@ -473,7 +473,7 @@ class NodeServiceTests {
     val amount = authEvent.data.amount
     val totalAmount = amount + fee
 
-    val feeEuro = EuroUtils.euroCentsToEuro(fee)
+    val feeEuro = EuroUtils.euroCentsToEuro(fee.toLong())
     val totalAmountEuro = EuroUtils.euroCentsToEuro(totalAmount)
 
     val feeEuroCents = BigDecimal(fee)
@@ -620,7 +620,7 @@ class NodeServiceTests {
     val amount = authEvent.data.amount
     val totalAmount = amount + fee
 
-    val feeEuro = EuroUtils.euroCentsToEuro(fee)
+    val feeEuro = EuroUtils.euroCentsToEuro(fee.toLong())
     val totalAmountEuro = EuroUtils.euroCentsToEuro(totalAmount)
 
     val feeEuroCents = BigDecimal(fee)
@@ -1215,7 +1215,7 @@ class NodeServiceTests {
     val amount = authEvent.data.amount
     val totalAmount = amount + fee
 
-    val feeEuro = EuroUtils.euroCentsToEuro(fee)
+    val feeEuro = EuroUtils.euroCentsToEuro(fee.toLong())
     val totalAmountEuro = EuroUtils.euroCentsToEuro(totalAmount)
 
     val feeEuroCents = BigDecimal(fee)
@@ -1340,7 +1340,7 @@ class NodeServiceTests {
     val amount = authEvent.data.amount
     val totalAmount = amount + fee
 
-    val feeEuro = EuroUtils.euroCentsToEuro(fee)
+    val feeEuro = EuroUtils.euroCentsToEuro(fee.toLong())
     val totalAmountEuro = EuroUtils.euroCentsToEuro(totalAmount)
 
     val feeEuroCents = BigDecimal(fee)
@@ -1696,7 +1696,7 @@ class NodeServiceTests {
     val amount = authEvent.data.amount
     val totalAmount = amount + fee
 
-    val feeEuro = EuroUtils.euroCentsToEuro(fee)
+    val feeEuro = EuroUtils.euroCentsToEuro(fee.toLong())
     val totalAmountEuro = EuroUtils.euroCentsToEuro(totalAmount)
 
     val feeEuroCents = BigDecimal(fee)
@@ -1845,7 +1845,7 @@ class NodeServiceTests {
     val amount = authEvent.data.amount
     val totalAmount = amount + fee
 
-    val feeEuro = EuroUtils.euroCentsToEuro(fee)
+    val feeEuro = EuroUtils.euroCentsToEuro(fee.toLong())
     val totalAmountEuro = EuroUtils.euroCentsToEuro(totalAmount)
 
     val feeEuroCents = BigDecimal(fee)
@@ -2239,7 +2239,7 @@ class NodeServiceTests {
     val amount = authEvent.data.amount
     val totalAmount = amount + fee
 
-    val feeEuro = EuroUtils.euroCentsToEuro(fee)
+    val feeEuro = EuroUtils.euroCentsToEuro(fee.toLong())
     val totalAmountEuro = EuroUtils.euroCentsToEuro(totalAmount)
 
     val feeEuroCents = BigDecimal(fee)
@@ -2385,7 +2385,7 @@ class NodeServiceTests {
     val amount = authEvent.data.amount
     val totalAmount = amount + fee
 
-    val feeEuro = EuroUtils.euroCentsToEuro(fee)
+    val feeEuro = EuroUtils.euroCentsToEuro(fee.toLong())
     val totalAmountEuro = EuroUtils.euroCentsToEuro(totalAmount)
 
     val feeEuroCents = BigDecimal(fee)
@@ -2778,7 +2778,7 @@ class NodeServiceTests {
     val amount = authEvent.data.amount
     val totalAmount = amount + fee
 
-    val feeEuro = EuroUtils.euroCentsToEuro(fee)
+    val feeEuro = EuroUtils.euroCentsToEuro(fee.toLong())
     val totalAmountEuro = EuroUtils.euroCentsToEuro(totalAmount)
 
     val feeEuroCents = BigDecimal(fee)
@@ -2924,7 +2924,7 @@ class NodeServiceTests {
     val amount = authEvent.data.amount
     val totalAmount = amount + fee
 
-    val feeEuro = EuroUtils.euroCentsToEuro(fee)
+    val feeEuro = EuroUtils.euroCentsToEuro(fee.toLong())
     val totalAmountEuro = EuroUtils.euroCentsToEuro(totalAmount)
 
     val feeEuroCents = BigDecimal(fee)
@@ -3317,7 +3317,7 @@ class NodeServiceTests {
     val amount = authEvent.data.amount
     val totalAmount = amount + fee
 
-    val feeEuro = EuroUtils.euroCentsToEuro(fee)
+    val feeEuro = EuroUtils.euroCentsToEuro(fee.toLong())
     val totalAmountEuro = EuroUtils.euroCentsToEuro(totalAmount)
 
     val feeEuroCents = BigDecimal(fee)
@@ -3468,7 +3468,7 @@ class NodeServiceTests {
     val amount = authEvent.data.amount
     val totalAmount = amount + fee
 
-    val feeEuro = EuroUtils.euroCentsToEuro(fee)
+    val feeEuro = EuroUtils.euroCentsToEuro(fee.toLong())
     val totalAmountEuro = EuroUtils.euroCentsToEuro(totalAmount)
 
     val feeEuroCents = BigDecimal(fee)
@@ -3866,7 +3866,7 @@ class NodeServiceTests {
     val amount = authEvent.data.amount
     val totalAmount = amount + fee
 
-    val feeEuro = EuroUtils.euroCentsToEuro(fee)
+    val feeEuro = EuroUtils.euroCentsToEuro(fee.toLong())
     val totalAmountEuro = EuroUtils.euroCentsToEuro(totalAmount)
 
     val feeEuroCents = BigDecimal(fee)
@@ -4020,7 +4020,7 @@ class NodeServiceTests {
     val amount = authEvent.data.amount
     val totalAmount = amount + fee
 
-    val feeEuro = EuroUtils.euroCentsToEuro(fee)
+    val feeEuro = EuroUtils.euroCentsToEuro(fee.toLong())
     val totalAmountEuro = EuroUtils.euroCentsToEuro(totalAmount)
 
     val feeEuroCents = BigDecimal(fee)
@@ -4420,7 +4420,7 @@ class NodeServiceTests {
     val amount = authEvent.data.amount
     val totalAmount = amount + fee
 
-    val feeEuro = EuroUtils.euroCentsToEuro(fee)
+    val feeEuro = EuroUtils.euroCentsToEuro(fee.toLong())
     val totalAmountEuro = EuroUtils.euroCentsToEuro(totalAmount)
 
     val feeEuroCents = BigDecimal(fee)
@@ -4570,7 +4570,7 @@ class NodeServiceTests {
     val amount = authEvent.data.amount
     val totalAmount = amount + fee
 
-    val feeEuro = EuroUtils.euroCentsToEuro(fee)
+    val feeEuro = EuroUtils.euroCentsToEuro(fee.toLong())
     val totalAmountEuro = EuroUtils.euroCentsToEuro(totalAmount)
 
     val feeEuroCents = BigDecimal(fee)
@@ -4965,7 +4965,7 @@ class NodeServiceTests {
     val amount = authEvent.data.amount
     val totalAmount = amount + fee
 
-    val feeEuro = EuroUtils.euroCentsToEuro(fee)
+    val feeEuro = EuroUtils.euroCentsToEuro(fee.toLong())
     val totalAmountEuro = EuroUtils.euroCentsToEuro(totalAmount)
 
     val feeEuroCents = BigDecimal(fee)
@@ -5358,7 +5358,7 @@ class NodeServiceTests {
     val amount = authEvent.data.amount
     val totalAmount = amount + fee
 
-    val feeEuro = EuroUtils.euroCentsToEuro(fee)
+    val feeEuro = EuroUtils.euroCentsToEuro(fee.toLong())
     val totalAmountEuro = EuroUtils.euroCentsToEuro(totalAmount)
 
     val feeEuroCents = BigDecimal(fee)
@@ -5739,7 +5739,7 @@ class NodeServiceTests {
     val amount = authEvent.data.amount
     val totalAmount = amount + fee
 
-    val feeEuro = EuroUtils.euroCentsToEuro(fee)
+    val feeEuro = EuroUtils.euroCentsToEuro(fee.toLong())
     val totalAmountEuro = EuroUtils.euroCentsToEuro(totalAmount)
 
     val feeEuroCents = BigDecimal(fee)
@@ -6104,7 +6104,7 @@ class NodeServiceTests {
     val amount = authEvent.data.amount
     val totalAmount = amount + fee
 
-    val feeEuro = EuroUtils.euroCentsToEuro(fee)
+    val feeEuro = EuroUtils.euroCentsToEuro(fee.toLong())
     val totalAmountEuro = EuroUtils.euroCentsToEuro(totalAmount)
 
     val feeEuroCents = BigDecimal(fee)
@@ -6512,7 +6512,7 @@ class NodeServiceTests {
     val amount = authEvent.data.amount
     val totalAmount = amount + fee
 
-    val feeEuro = EuroUtils.euroCentsToEuro(fee)
+    val feeEuro = EuroUtils.euroCentsToEuro(fee.toLong())
     val totalAmountEuro = EuroUtils.euroCentsToEuro(totalAmount)
 
     val feeEuroCents = BigDecimal(fee)
@@ -6921,7 +6921,7 @@ class NodeServiceTests {
     val amount = authEvent.data.amount
     val totalAmount = amount + fee
 
-    val feeEuro = EuroUtils.euroCentsToEuro(fee)
+    val feeEuro = EuroUtils.euroCentsToEuro(fee.toLong())
     val totalAmountEuro = EuroUtils.euroCentsToEuro(totalAmount)
 
     val feeEuroCents = BigDecimal(fee)
@@ -7324,7 +7324,7 @@ class NodeServiceTests {
     val amount = authEvent.data.amount
     val totalAmount = amount + fee
 
-    val feeEuro = EuroUtils.euroCentsToEuro(fee)
+    val feeEuro = EuroUtils.euroCentsToEuro(fee.toLong())
     val totalAmountEuro = EuroUtils.euroCentsToEuro(totalAmount)
 
     val feeEuroCents = BigDecimal(fee)
@@ -7871,7 +7871,7 @@ class NodeServiceTests {
       val amount = authEvent.data.amount
       val totalAmount = amount + fee
 
-      val feeEuro = EuroUtils.euroCentsToEuro(fee)
+      val feeEuro = EuroUtils.euroCentsToEuro(fee.toLong())
       val totalAmountEuro = EuroUtils.euroCentsToEuro(totalAmount)
 
       val feeEuroCents = BigDecimal(fee)
@@ -8026,7 +8026,7 @@ class NodeServiceTests {
       val amount = authEvent.data.amount
       val totalAmount = amount + fee
 
-      val feeEuro = EuroUtils.euroCentsToEuro(fee)
+      val feeEuro = EuroUtils.euroCentsToEuro(fee.toLong())
       val totalAmountEuro = EuroUtils.euroCentsToEuro(totalAmount)
 
       val feeEuroCents = BigDecimal(fee)
